Upgrade to React Native 0.73.4 (see details below)
Upgrade Procedure
*****************

Since the app was very outdated (RN 0.64.0, released in May 2021), I decided to create a new project
in another directory with 'npx react-native@latest init MemoryGame'.

After this you need to do 'cd ios' and 'bundle install' (only once ever), which creates the
Gemfile.lock. Then 'bundle exec pod install' which creates the Podfile.lock.

Then I upgraded the packages in the template. Prettier with 'yarn add --dev --exact prettier',
TypeScript of to latest (yarn upgrade --latest typescript), and also @types/react and
@types/react-native.

Changed the Prettier config in .prettierrc.js to be the same than in the existing project to have
the same formatting. The config option jsxBracketSameLine is now called bracketSameLine.

Then I added the same packages I have on the existing project:

$ yarn add mobx mobx-react-lite

$ yarn add -E react-native-vector-icons
$ yarn add -D -E @types/react-native-vector-icons
Also did the setup in android/app/build.gradle and Xcode as explained in
https://github.com/oblador/react-native-vector-icons#installation

Reanimated installation and setup following
https://docs.swmansion.com/react-native-reanimated/docs/fundamentals/getting-started/

$ yarn add react-native-linear-gradient

Note that I had to do 'bundle exec pod install' for some libraries, which adds the library to
Podfile.lock.

When all packages were installed and the important ones were at the latest version, I copied all the
code in 'src', the README.md and the App.tsx files of the existing project to the new project.

After verifying that the new app worked on Android and iOS, then I copied all the contents of the new
project into the existing project.

I had to delete some files too, like .buckconfig, .gitattributes, _editorconfig etc. I used `diff` to
see the differences.

After verifying that the existing project also worked I did this commit.

Note that there is something to fix: the default color of the text in Android was black on RN 0.64
but is white on RN 0.73.