Upgrade to React Native 0.75.4 (see details below)
Procedure is the same than the upgrade to 0.73.4 in f8a83e035b0405a53be86e3f2c344510f6b9b3fc.

Create new project:

$ npx @react-native-community/cli@latest init MemoryGame

Then do:

$ cd ios
$ bundle install
$ bundle exec pod install

Note: the 'bundle exec pod install' failed because I don't have the latest Xcode:

  React Native requires XCode >= 14.3. Found 14.2.
  [!] Invalid `Podfile` file: Please upgrade XCode.

Then upgrade and install the packages:

$ yarn up -E typescript

$ yarn up -E prettier

Copy the prettier config file to the new project.

$ yarn add -E mobx mobx-react-lite

$ yarn up -E @types/react

$ yarn add -E react-native-vector-icons

I did the Android setup in https://github.com/oblador/react-native-vector-icons, but not the iOS
because I can't run 'pod install'.

$ yarn add -D -E @types/react-native-vector-icons

$ yarn add -E react-native-reanimated

Did the setup in https://docs.swmansion.com/react-native-reanimated/docs/fundamentals/getting-started/
but again did not do the 'pod install'.

After all this, I copied the code in 'src' in the existing project to the new one.

I then formatted the files with prettier.

After verifying that the new app worked on Android (not iOS because I can't 'pod install), then I
copied the contents of the new project into the existing project, overriding folders and deleting
the files that were not present in the new project.

Note that there are some files like the README.md that don't need to be copied.

Also note that there's no need to use `diff` in the procedure.

After copying, verify that the existing project works and the commit.