Fix reanimated warning "Writing to `value` during component render"
The full warning is:

[Reanimated] Writing to `value` during component render. Please ensure that you do not access the `value`
property or use `set` method of a shared value while React is rendering a component.
If you don't want to see this message, you can disable the `strict` mode. Refer to:
https://docs.swmansion.com/react-native-reanimated/docs/debugging/logger-configuration for more details.

This warning was introduced in version 3.16, in this commit:
https://github.com/software-mansion/react-native-reanimated/pull/6310

Also see this issue: https://github.com/dohooo/react-native-reanimated-carousel/issues/706